Made the expiration date for the policy a utc date. (As advised by kookster:
"
I also posted this on the elcgit parent of this fork, but you seem to be making updates more often, so thought you might incorporate it sooner.

If you have used newish RoR ability to set a time zone, then you need to make sure the expiration_date included in the policy block for s3 is utc.

Line 30:

    expiration_date = 1.hours.from_now.strftime('%Y-%m-%dT%H:%M:%S.000Z')

Should be:

    expiration_date = 1.hours.from_now.utc.strftime('%Y-%m-%dT%H:%M:%S.000Z')
"
)